move rpc models out of core
