+++ applications/minotari_node/src/http/handler/get_tip_info.rs
@@ -11,9 +11,10 @@ use axum::{
 };
 use log::debug;
 use tari_core::{
-    base_node::rpc::{models::TipInfoResponse, query_service, BaseNodeWalletQueryService},
+    base_node::rpc::{query_service, BaseNodeWalletQueryService},
     chain_storage::BlockchainBackend,
 };
+use tari_transaction_components::rpc::models::TipInfoResponse;
 use tonic::service::AxumBody;
 
 use crate::http::handler::{error_handler_with_message, ErrorResponse};
