+++ base_layer/core/src/base_node/rpc/mod.rs
@@ -6,7 +6,6 @@ mod service;
 pub mod sync_utxos_by_block_task;
 #[cfg(feature = "base_node")]
 pub use service::BaseNodeWalletRpcService;
-pub mod models;
 
 #[cfg(feature = "base_node")]
 pub mod query_service;
@@ -15,6 +14,7 @@ use std::{error::Error, fmt::Debug};
 
 use tari_comms::protocol::rpc::{Request, Response, RpcStatus, Streaming};
 use tari_comms_rpc_macros::tari_rpc;
+use tari_transaction_components::rpc::models;
 #[cfg(feature = "base_node")]
 use url::Url;
