+++ base_layer/core/src/blocks/block_header.rs
@@ -300,6 +300,53 @@ impl From<NewBlockHeaderTemplate> for BlockHeader {
     }
 }
 
+impl From<tari_transaction_components::rpc::models::BlockHeader> for BlockHeader {
+    fn from(header: tari_transaction_components::rpc::models::BlockHeader) -> Self {
+        Self {
+            version: header.version,
+            height: header.height,
+            prev_hash: header.prev_hash,
+            timestamp: header.timestamp,
+            input_mr: header.input_mr,
+            output_mr: header.output_mr,
+            block_output_mr: header.block_output_mr,
+            output_smt_size: header.output_smt_size,
+            kernel_mr: header.kernel_mr,
+            kernel_mmr_size: header.kernel_mmr_size,
+            total_kernel_offset: header.total_kernel_offset,
+            total_script_offset: header.total_script_offset,
+            validator_node_mr: header.validator_node_mr,
+            validator_node_size: header.validator_node_size,
+            pow: header.pow,
+            nonce: header.nonce,
+        }
+    }
+}
+
+impl From<BlockHeader> for tari_transaction_components::rpc::models::BlockHeader {
+    fn from(header: BlockHeader) -> Self {
+        Self {
+            hash: header.hash(),
+            version: header.version,
+            height: header.height,
+            prev_hash: header.prev_hash,
+            timestamp: header.timestamp,
+            input_mr: header.input_mr,
+            output_mr: header.output_mr,
+            block_output_mr: header.block_output_mr,
+            output_smt_size: header.output_smt_size,
+            kernel_mr: header.kernel_mr,
+            kernel_mmr_size: header.kernel_mmr_size,
+            total_kernel_offset: header.total_kernel_offset,
+            total_script_offset: header.total_script_offset,
+            validator_node_mr: header.validator_node_mr,
+            validator_node_size: header.validator_node_size,
+            pow: header.pow,
+            nonce: header.nonce,
+        }
+    }
+}
+
 impl PartialEq for BlockHeader {
     fn eq(&self, other: &Self) -> bool {
         self.hash() == other.hash()
