+++ base_layer/core/src/mempool/unconfirmed_pool/unconfirmed_pool.rs
@@ -29,6 +29,7 @@ use log::*;
 use serde::{Deserialize, Serialize};
 use tari_common_types::types::{CompressedSignature, FixedHash, HashOutput, PrivateKey};
 use tari_transaction_components::{
+    rpc::models::FeePerGramStat,
     tari_amount::MicroMinotari,
     transaction_components::{Transaction, TransactionError},
     weight::TransactionWeight,
@@ -41,11 +42,9 @@ use crate::{
         priority::{FeePriority, PrioritizedTransaction},
         shrink_hashmap::shrink_hashmap,
         unconfirmed_pool::UnconfirmedPoolError,
-        FeePerGramStat,
         MempoolError,
     },
 };
-
 pub const LOG_TARGET: &str = "c::mp::unconfirmed_pool::unconfirmed_pool_storage";
 
 pub type TransactionKey = usize;
