+++ base_layer/transaction_components/src/rpc/mod.rs
@@ -1,0 +1,4 @@
+// Copyright 2025 The Tari Project
+// SPDX-License-Identifier: BSD-3-Clause
+
+pub mod models;
