+++ base_layer/transaction_components/src/rpc/models/get_utxos_deleted_info.rs
