+++ base_layer/transaction_components/src/rpc/models/mod.rs
@@ -1,6 +1,7 @@
 // Copyright 2025 The Tari Project
 // SPDX-License-Identifier: BSD-3-Clause
 
+mod fee_per_gram;
 mod get_header_by_height;
 mod get_tip_info;
 mod get_utxos_by_block;
@@ -10,6 +11,7 @@ mod sync_utxos_by_block;
 mod transaction_query;
 mod tx_submission_response;
 
+pub use fee_per_gram::*;
 pub use get_header_by_height::*;
 pub use get_tip_info::*;
 pub use get_utxos_by_block::*;
