+++ base_layer/transaction_components/src/rpc/models/sync_utxos_by_block.rs
