+++ base_layer/transaction_components/src/validation/error.rs
@@ -31,57 +31,23 @@ use crate::{
 pub enum AggregatedBodyValidationError {
     #[error("Serialization failed: {0}")]
     SerializationError(String),
-    // #[error("Block header validation failed: {0}")]
-    // BlockHeaderError(#[from] BlockHeaderValidationError),
-    // #[error("Block validation error: {0}")]
-    // BlockError(#[from] BlockValidationError),
     #[error("Contains kernels or inputs that are not yet spendable")]
     MaturityError,
     #[error("The block weight ({actual_weight}) is above the maximum ({max_weight})")]
     BlockTooLarge { actual_weight: u64, max_weight: u64 },
-    // #[error("Contains {} unknown inputs", .0.len())]
-    // UnknownInputs(Vec<HashOutput>),
-    // #[error("Contains an unknown input")]
-    // UnknownInput,
     #[error("The transaction is invalid: {0}")]
     TransactionError(#[from] TransactionError),
-    // #[error("Fatal storage error during validation: {0}")]
-    // FatalStorageError(String),
     #[error(
         "The total expected supply plus the total accumulated (offset) excess does not equal the sum of all UTXO \
          commitments."
     )]
     InvalidAccountingBalance,
-    // #[error("Transaction contains already spent inputs")]
-    // ContainsSTxO,
-    // #[error("Transaction contains outputs that already exist")]
-    // ContainsTxO,
-    // #[error("Transaction contains an output commitment that already exists")]
-    // ContainsDuplicateUtxoCommitment,
-    // #[error("Final state validation failed: The UTXO set did not balance with the expected emission at height
-    // {0}")] ChainBalanceValidationFailed(u64),
-    // #[error("The total value + fees of the block exceeds the maximum allowance on chain")]
-    // CoinbaseExceedsMaxLimit,
-    // #[error("Proof of work error: {0}")]
-    // ProofOfWorkError(#[from] PowError),
-    // #[error("Attempted to validate genesis block")]
-    // ValidatingGenesis,
     #[error("Duplicate or unsorted input found in block body")]
     UnsortedOrDuplicateInput,
     #[error("Duplicate or unsorted output found in block body")]
     UnsortedOrDuplicateOutput,
     #[error("Duplicate or unsorted kernel found in block body")]
     UnsortedOrDuplicateKernel,
-    // #[error("Error in merge mine data:{0}")]
-    // MergeMineError(#[from] MergeMineError),
-    // #[error("Maximum transaction weight exceeded")]
-    // MaxTransactionWeightExceeded,
-    // #[error("Expected block height to be {expected}, but was {block_height}")]
-    // IncorrectHeight { expected: u64, block_height: u64 },
-    // #[error("Expected block previous hash to be {expected}, but was {block_hash}")]
-    // IncorrectPreviousHash { expected: String, block_hash: String },
-    // #[error("Bad block with hash '{hash}' and reason '{reason}' found")]
-    // BadBlockFound { hash: String, reason: String },
     #[error(
         "Script exceeded maximum script size, expected less than {max_script_size} but was
     {actual_script_size}"
@@ -100,12 +66,8 @@ pub enum AggregatedBodyValidationError {
     },
     #[error("Consensus Error: {0}")]
     ConsensusError(String),
-    // #[error("Duplicate kernel Error: {0}")]
-    // DuplicateKernelError(String),
     #[error("Covenant failed to validate: {0}")]
     CovenantError(#[from] CovenantError),
-    // #[error("Invalid or unsupported blockchain version {version}")]
-    // InvalidBlockchainVersion { version: u16 },
     #[error("Contains Invalid Burn: {0}")]
     InvalidBurnError(String),
     #[error("Output type '{output_type}' is not permitted")]
@@ -118,49 +80,16 @@ pub enum AggregatedBodyValidationError {
     ValidatorNodeRegistrationMinDepositAmount { min: MicroMinotari, actual: MicroMinotari },
     #[error("Validator registration has invalid maturity {actual}, must be at least {min}")]
     ValidatorNodeRegistrationMinLockHeight { min: u64, actual: u64 },
-    // #[error("Sidechain ID knowledge proof not valid for template registration")]
-    // TemplateInvalidSidechainIdKnowledgeProof,
     #[error("Author signature not valid for template registration")]
     TemplateAuthorSignatureNotValid,
-    // #[error("Sidechain ID knowledge proof not valid for confidential output")]
-    // ConfidentialOutputSidechainIdKnowledgeProofNotValid,
     #[error("Validator node registration signature failed verification")]
     InvalidValidatorNodeSignature,
     #[error("Sidechain ID knowledge proof not valid for validator node registration")]
     ValidatorNodeInvalidSidechainIdKnowledgeProof,
-    // #[error(
-    //     "An unexpected number of timestamps were provided to the header validator. THIS IS A BUG. Expected \
-    //      {expected}, got {actual}"
-    // )]
-    // IncorrectNumberOfTimestampsProvided { expected: u64, actual: u64 },
-    // #[error("Invalid difficulty: {0}")]
-    // DifficultyError(#[from] DifficultyError),
     #[error("Covenant too large. Max size: {max_size}, Actual size: {actual_size}")]
     CovenantTooLarge { max_size: usize, actual_size: usize },
     #[error("Invalid Serialized Public key: {0}")]
     InvalidSerializedPublicKey(String),
-    // #[error("Sidechain proof invalid: `{0}`")]
-    // SidechainProofInvalid(#[from] SidechainProofValidationError),
-    // #[error("Sidechain eviction proof submitted for unregistered validator {validator_pk}")]
-    // SidechainEvictionProofValidatorNotFound { validator_pk: String },
-    // #[error(
-    //     "Sidechain eviction proof invalid: given epoch {epoch} is greater than the epoch at tip height
-    // {tip_height}" )]
-    // SidechainEvictionProofInvalidEpoch { epoch: VnEpoch, tip_height: u64 },
-    // #[error("Validator node already registered: {public_key}")]
-    // ValidatorNodeAlreadyRegistered { public_key: String },
-    // #[error("Validator node {public_key} not registered: {details}")]
-    // ValidatorNodeNotRegistered { public_key: String, details: String },
-    // #[error("Validator registration {public_key} invalid: max epoch {max_epoch} < current epoch {current_epoch}")]
-    // ValidatorNodeRegistrationMaxEpoch {
-    //     public_key: String,
-    //     current_epoch: VnEpoch,
-    //     max_epoch: VnEpoch,
-    // },
-    // #[error("{output_type} output rule disallows the spend: {details}")]
-    // OutputSpendRuleDisallow { output_type: OutputType, details: String },
-    // #[error("Output type '{output_type}' does not match sidechain data")]
-    // OutputTypeNotMatchSidechainData { output_type: OutputType, details: String },
 }
 
 impl From<ByteArrayError> for AggregatedBodyValidationError {
@@ -168,67 +97,3 @@ impl From<ByteArrayError> for AggregatedBodyValidationError {
         Self::InvalidSerializedPublicKey(err.to_string())
     }
 }
-
-// impl AggregatedBodyValidationError {
-//     pub fn get_ban_reason(&self) -> Option<BanReason> {
-//         match self {
-//         //     ValidationError::ProofOfWorkError(e) => e.get_ban_reason(),
-//             err @ AggregatedBodyValidationError::SerializationError(_) |
-//         //     err @ ValidationError::BlockHeaderError(_) |
-//         //     err @ ValidationError::BlockError(_) |
-//         //     err @ ValidationError::MaturityError |
-//         //     err @ ValidationError::BlockTooLarge { .. } |
-//         //     err @ ValidationError::UnknownInputs(_) |
-//         //     err @ ValidationError::UnknownInput |
-//         //     err @ ValidationError::TransactionError(_) |
-//         //     err @ ValidationError::InvalidAccountingBalance |
-//         //     err @ ValidationError::ContainsSTxO |
-//         //     err @ ValidationError::ContainsTxO |
-//         //     err @ ValidationError::ContainsDuplicateUtxoCommitment |
-//         //     err @ ValidationError::ChainBalanceValidationFailed(_) |
-//         //     err @ ValidationError::ValidatingGenesis |
-//         //     err @ ValidationError::UnsortedOrDuplicateInput |
-//         //     err @ ValidationError::UnsortedOrDuplicateOutput |
-//         //     err @ ValidationError::UnsortedOrDuplicateKernel |
-//         //     err @ ValidationError::MaxTransactionWeightExceeded |
-//         //     err @ ValidationError::IncorrectHeight { .. } |
-//         //     err @ ValidationError::IncorrectPreviousHash { .. } |
-//         //     err @ ValidationError::BadBlockFound { .. } |
-//             err @ AggregatedBodyValidationError::TariScriptExceedsMaxSize { .. } |
-//             err @ AggregatedBodyValidationError::EncryptedDataExceedsMaxSize { .. } |
-//             err @ AggregatedBodyValidationError::ConsensusError(_) |
-//         //     err @ ValidationError::DuplicateKernelError(_) |
-//         //     err @ ValidationError::CovenantError(_) |
-//         //     err @ ValidationError::InvalidBlockchainVersion { .. } |
-//         //     err @ ValidationError::InvalidBurnError(_) |
-//             err @ AggregatedBodyValidationError::OutputTypeNotPermitted { .. } |
-//             err @ AggregatedBodyValidationError::RangeProofTypeNotPermitted { .. } |
-//             err @ AggregatedBodyValidationError::OutputTypeNotMatchedToRangeProofType { .. } |
-//         //     err @ ValidationError::ValidatorNodeRegistrationMinDepositAmount { .. } |
-//         //     err @ ValidationError::ValidatorNodeRegistrationMinLockHeight { .. } |
-//         //     err @ ValidationError::InvalidValidatorNodeSignature |
-//         //     err @ ValidationError::ValidatorNodeInvalidSidechainIdKnowledgeProof |
-//         //     err @ ValidationError::TemplateInvalidSidechainIdKnowledgeProof |
-//         //     err @ ValidationError::TemplateAuthorSignatureNotValid |
-//         //     err @ ValidationError::ConfidentialOutputSidechainIdKnowledgeProofNotValid |
-//         //     err @ ValidationError::DifficultyError(_) |
-//         //     err @ ValidationError::CoinbaseExceedsMaxLimit |
-//             err @ AggregatedBodyValidationError::CovenantTooLarge { .. } |
-//         //     err @ ValidationError::InvalidSerializedPublicKey(_) |
-//         //     err @ ValidationError::SidechainEvictionProofValidatorNotFound { .. } |
-//         //     err @ ValidationError::SidechainProofInvalid(_) |
-//         //     err @ ValidationError::SidechainEvictionProofInvalidEpoch { .. } |
-//         //     err @ ValidationError::ValidatorNodeAlreadyRegistered { .. } |
-//         //     err @ ValidationError::ValidatorNodeNotRegistered { .. } |
-//         //     err @ ValidationError::ValidatorNodeRegistrationMaxEpoch { .. } |
-//         //     err @ ValidationError::OutputTypeNotMatchSidechainData { .. } |
-//         //     err @ ValidationError::OutputSpendRuleDisallow { .. } =>
-//             => Some(BanReason {
-//                 reason: err.to_string(),
-//                 ban_duration: BanPeriod::Long,
-//             }),
-//         //     ValidationError::MergeMineError(e) => e.get_ban_reason(),
-//         //     ValidationError::FatalStorageError(_) | ValidationError::IncorrectNumberOfTimestampsProvided { .. } =>
-//         // None, }
-//     }
-// }
